Ryddet litt mer i tester og strukturert ting i egne filer. Dratt gjennom en behandling og fikset noen bugs.
