+++ src/frontend/context/FritekstVedtakBegrunnelserContext.tsx
@@ -136,6 +136,9 @@ const [FritekstVedtakBegrunnelserProvider, useFritekstVedtakBegrunnelser] = cons
         };
 
         const onSubmit = () => {
+            if (!vedtaksperiode.periodeFom) {
+                throw new Error('Prøver å legge til en begrunnelse på en periode uten fom');
+            }
             settFritekstSubmit(FritekstSubmit.POST);
 
             if (fagsak.status === RessursStatus.SUKSESS) {
