+++ src/frontend/komponenter/Fagsak/Simulering/Simulering.tsx
@@ -41,6 +41,7 @@ const Simulering: React.FunctionComponent<ISimuleringProps> = ({ åpenBehandling
     const { settFagsak } = useFagsakRessurser();
 
     const nesteOnClick = () => {
+        history.push(`/fagsak/${fagsak.id}/${åpenBehandling?.behandlingId}/vedtak`);
         onSubmit<ITilbakekreving | undefined>(
             {
                 data: hentSkjemadata(),
