+++ src/frontend/utils/kalender/index.ts
@@ -6,3 +6,4 @@ export * from './periode';
 export * from './kalenderMåned';
 export * from './utils';
 export * from './aritmetikk';
+export * from './sammenligning';
