+++ src/frontend/utils/kalender/kalender.ts
@@ -55,36 +55,3 @@ export const kalenderDatoMedFallback = (
 ): DagMånedÅr => (dato ? parseIso8601String(dato) : fallbackDato);
 
 export const kalenderDato = (dato: FamilieIsoDate): DagMånedÅr => parseIso8601String(dato);
-
-// Sammenligningsfunksjoner
-export const erISammeMåned = (dato1: DagMånedÅr, dato2: DagMånedÅr) => {
-    return dato1.måned === dato2.måned && dato1.år === dato2.år;
-};
-
-export const erFør = (dato1: DagMånedÅr, dato2: DagMånedÅr) => {
-    if (dato1.dag < dato2.dag && dato1.måned <= dato2.måned && dato1.år <= dato2.år) {
-        return true;
-    }
-
-    if (dato1.måned < dato2.måned && dato1.år <= dato2.år) {
-        return true;
-    }
-
-    return dato1.år < dato2.år;
-};
-
-export const erEtter = (dato1: DagMånedÅr, dato2: DagMånedÅr) => {
-    if (dato1.dag > dato2.dag && dato1.måned >= dato2.måned && dato1.år >= dato2.år) {
-        return true;
-    }
-
-    if (dato1.måned > dato2.måned && dato1.år >= dato2.år) {
-        return true;
-    }
-
-    return dato1.år > dato2.år;
-};
-
-export const erSamme = (dato1: DagMånedÅr, dato2: DagMånedÅr) => {
-    return dato1.dag === dato2.dag && dato1.måned === dato2.måned && dato1.år === dato2.år;
-};
